+++ test/test-operation.js
@@ -398,6 +398,24 @@ function runTests (mode) {
             })
             .then(done, done);
         });
+
+        it('should not return an INVALID_CONENT_TYPE error for empty body (Issue 164)', function (done) {
+          var cSwaggerDoc = _.cloneDeep(helpers.swaggerDoc);
+
+          cSwaggerDoc.paths['/user'].post.parameters[0].required = false;
+          cSwaggerDoc.paths['/user'].post.consumes = ['application/json'];
+
+          Sway.create({
+            definition: cSwaggerDoc
+          })
+            .then(function (api) {
+              var results = api.getOperation('/user', 'post').validateRequest({});
+
+              assert.equal(results.warnings.length, 0);
+              assert.equal(results.errors.length, 0);
+            })
+            .then(done, done);
+        });
       });
 
       describe('validate parameters', function () {
@@ -587,6 +605,44 @@ function runTests (mode) {
             assert.equal(scenario[1], err.message);
           }
         });
+      });
+
+      it('should not return an INVALID_CONENT_TYPE error for empty body (Issue 164)', function (done) {
+        var cSwaggerDoc = _.cloneDeep(tHelpers.swaggerDoc);
+
+        cSwaggerDoc.paths['/user'].post.produces = ['application/xml'];
+        cSwaggerDoc.paths['/user'].post.responses.default.schema = {
+          type: 'object'
+        };
+
+        Sway.create({
+          definition: cSwaggerDoc
+        })
+          .then(function (api) {
+            var results = api.getOperation('/user', 'post').validateResponse({
+              headers: {
+                'Content-Type': 'application/json'
+              }
+            });
+
+            assert.equal(results.warnings.length, 0);
+            assert.deepEqual(results.errors, [
+              {
+                code: 'INVALID_RESPONSE_BODY',
+                errors: [
+                  {
+                    code: 'INVALID_TYPE',
+                    params: ['object', 'undefined'],
+                    message: 'Expected type object but found type undefined',
+                    path: []
+                  }
+                ],
+                message: 'Invalid body: Expected type object but found type undefined',
+                path: []
+              }
+            ]);
+          })
+          .then(done, done);
       });
 
       // We only test that Operation#validateResponse handles missing responses because the testing of the remainder